hello-world/pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.13",
    "langchain-core>=1.2.22",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "python-dotenv>=1.2.2",
]
